requirements.txt
fastapi
uvicorn
pydantic
pandas
google-genai
python-dotenv
psycopg2-binary